+++ R/tidying_functions.R
@@ -194,3 +194,62 @@ norm_standardize <- function(data, scales, angles = octants(), instrument,
     as.data.frame(scores) 
   }
 }
+
+#' Standardize circumplex scales using sample data
+#'
+#' Take in a data frame containing circumplex scales (or items) and return that
+#' same data frame with each specified variable transformed into standard scores
+#' (i.e., z-scores) based on observed means and SDs.
+#'
+#' @param data Required. A data frame or matrix containing at least circumplex
+#'   scales.
+#' @param scales Required. A character vector containing the column names, or a
+#'   numeric vector containing the column indexes, for the variables (scale
+#'   scores) to be standardized.
+#' @param na.rm Optional. A logical that determines whether to remove missing
+#'   values from scales when calculating the means and SDs used for
+#'   standardization (default = TRUE).
+#' @param prefix Optional. A string to include at the beginning of the newly
+#'   calculated scale variables' names, before the scale name and `suffix`
+#'   (default = "").
+#' @param suffix Optional. A string to include at the end of the newly
+#'   calculated scale variables' names, after the scale name and `prefix`
+#'   (default = "_z").
+#' @param append Optional. A logical that determines whether the calculated
+#'   standardized scores should be added as columns to `data` in the output or
+#'   the standardized scores alone should be output (default = TRUE).
+#' @return A data frame that contains the self-standardized versions of
+#'   `scales`.
+#' @export
+#' @family tidying functions
+#' @examples
+#' self_standardize(aw2009, scales = 1:8)
+self_standardize <- function(data, scales, na.rm = TRUE,
+                             prefix = "", suffix = "_z", append = TRUE) {
+  
+  stopifnot(is.data.frame(data) || is.matrix(data))
+  stopifnot(is_var(scales))
+  stopifnot(is_flag(na.rm))
+  stopifnot(is_char(prefix, n = 1))
+  stopifnot(is_char(suffix, n = 1))
+  stopifnot(is_flag(append))
+  
+  scale_data <- data[scales]
+  scale_names <- colnames(scale_data)
+  
+  scores <- matrix(NA, nrow = nrow(scale_data), ncol = length(scales))
+  colnames(scores) <- paste0(prefix, scale_names, suffix)
+  
+  for (i in 1:length(scales)) {
+    m_i <- mean(scale_data[[i]], na.rm = na.rm)
+    s_i <- sd(scale_data[[i]], na.rm = na.rm)
+    scores[, i] <- (scale_data[[i]] - m_i) / s_i
+  }
+  scores[is.nan(scores)] <- NA_real_
+  
+  if (append) {
+    cbind(data, scores)
+  } else {
+    as.data.frame(scores) 
+  }
+}