+++ NEWS.md
@@ -1,3 +1,5 @@
+# circumplex (development version)
+
 # circumplex 1.0.0
 
 ## Breaking changes